requirements.txt
airtable-python-wrapper==0.8.0
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.6
psycopg2==2.8.3
pytz==2017.3
requests==2.20.0
six==1.11.0
urllib3==1.24.2
whitenoise==4.1.3
